+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log
 
 
+## Unreleased
+
+**Implemented enhancements:**
+
+- Exclude .ignore langages from the "New Scratch" action [\#448](https://github.com/hsz/idea-gitignore/issues/#448)
+
+
 ## [v2.2.1](https://github.com/hsz/idea-gitignore/tree/v2.2.1) (2017-09-14)
 
 [Full Changelog](https://github.com/hsz/idea-gitignore/compare/v2.2.0...v2.2.1)
@@ -11,6 +18,7 @@
 - Write-unsafe context [\#471](https://github.com/hsz/idea-gitignore/issues/#471)
 - Error on opening .gitignore file [\#470](https://github.com/hsz/idea-gitignore/issues/#470)
 - File colouring not working [\#462](https://github.com/hsz/idea-gitignore/issues/#462)
+
 
 ## [v2.2.0](https://github.com/hsz/idea-gitignore/tree/v2.2.0) (2017-09-06)